Sagarnew (#8)
* Connected register, login and logout.

Changed approach for logging out. Not killing process anymore. Saving state in the ets table as :online/:offline and in the client as 3rd parameter of state.

* Corrections

* Tweet simulations, Changes in Mentions

Commented the old regex part for scanning mentions and added scan for string starting with "@User" instead.

* Simulating query and retweeting.

Call is timing out for large number of users.